# server/docs/swagger.yaml
basePath: /api
definitions:
  github_com_ai-comic-generator_server_internal_common.BaseResponse:
    properties:
      code:
        description: 业务状态码，0 表示成功，非 0 表示失败
        example: 0
        type: integer
      data:
        description: 响应数据，成功时返回；失败时通常为 null
      message:
        description: 响应消息，成功时为 ok，失败时为错误描述
        example: ok
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.ComicCharacter:
    properties:
      appearance:
        description: 外貌详细描述（供分镜与定妆照）
        type: string
      avatarUrl:
        description: 定妆照 URL（角色设定后生成）
        type: string
      name:
        description: 角色名称
        type: string
      personality:
        description: 性格特征
        type: string
      role:
        description: 角色类型：protagonist 主角 / antagonist 反派 / supporting 配角
        type: string
      visualAnchor:
        description: 外貌短锚点（15-40 字，每格生图强制复用）
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.ComicInfo:
    properties:
      captionTextMode:
        description: 文案模式：none / top / bubble
        type: string
      characters:
        description: 角色列表（已解析）
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ComicCharacter'
        type: array
      completedTime:
        description: 完成时间
        type: string
      composedLayout:
        allOf:
        - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ComposedLayoutResult'
        description: 排版结果（已解析）
      coverImage:
        description: 封面图 URL
        type: string
      createTime:
        description: 创建时间
        type: string
      errorMessage:
        description: 错误信息
        type: string
      id:
        description: 主键 ID
        type: integer
      imageBackend:
        description: 生图后端
        type: string
      panelCount:
        description: 分镜格数
        type: integer
      panelImages:
        description: 分镜图片列表（已解析）
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PanelImageResult'
        type: array
      phase:
        description: 当前阶段
        type: string
      publishResult:
        allOf:
        - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PublishResult'
        description: 发布结果（已解析）
      status:
        description: 任务总状态
        type: string
      storyIdeation:
        allOf:
        - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.StoryIdeationResult'
        description: 故事构思（已解析）
      storyboard:
        allOf:
        - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.StoryboardResult'
        description: 分镜脚本（已解析）
      style:
        description: 漫画风格
        type: string
      taskId:
        description: 任务 UUID
        type: string
      title:
        description: 漫画标题
        type: string
      titleOptions:
        allOf:
        - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.TitleOptionsResult'
        description: 标题推荐列表（已解析）
      topic:
        description: 创作主题
        type: string
      userAccount:
        description: 创建者账号（联查，不落库）
        type: string
      userDescription:
        description: 用户补充描述
        type: string
      userId:
        description: 所属用户 ID
        type: integer
      userName:
        description: 创建者昵称（联查，不落库）
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.ComicPageResult:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
      records:
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ComicInfo'
        type: array
      total:
        type: integer
    type: object
  github_com_ai-comic-generator_server_internal_model.ComposedLayoutResult:
    properties:
      assetUrls:
        description: 各页或各资源的最终文件 URL 列表
        items:
          type: string
        type: array
      coverImage:
        description: 封面图 URL（用于列表展示）
        type: string
      format:
        description: 排版格式：long_image 长图 / grid 宫格 / page_pdf 分页 PDF
        example: long_image
        type: string
      previewUrl:
        description: 合成预览图 URL
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.ConfirmStoryboardRequest:
    properties:
      storyboard:
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.StoryboardPanel'
        type: array
      taskId:
        type: string
    required:
    - storyboard
    - taskId
    type: object
  github_com_ai-comic-generator_server_internal_model.ConfirmTitleRequest:
    properties:
      taskId:
        type: string
      title:
        type: string
    required:
    - taskId
    - title
    type: object
  github_com_ai-comic-generator_server_internal_model.CreateComicRequest:
    properties:
      captionTextMode:
        enum:
        - none
        - top
        - bubble
        example: top
        type: string
      imageBackend:
        enum:
        - hunyuan
        - openai_image_1k
        - openai_image_4k
        example: hunyuan
        type: string
      panelCount:
        description: 分镜格数：4 / 6 / 8
        example: 4
        type: integer
      style:
        enum:
        - cartoon
        - realistic
        - chibi
        - animal
        example: cartoon
        type: string
      topic:
        example: 哪吒闹海
        type: string
      userDescription:
        example: 搞笑风格，保持角色一致性
        type: string
    required:
    - topic
    type: object
...
  github_com_ai-comic-generator_server_internal_model.PanelImageResult:
    properties:
      imagePrompt:
        description: 实际使用的生图 Prompt（便于追溯）
        type: string
      method:
        description: 图片来源：AI 生成 或 用户上传
        enum:
        - AI_GENERATE
        - UPLOAD
        type: string
      panelNo:
        description: 对应分镜格序号
        type: integer
      url:
        description: 图片访问地址
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.PublishComicRequest:
    properties:
      platform:
        enum:
        - WECHAT_MP
        example: WECHAT_MP
        type: string
      taskId:
        type: string
    required:
    - taskId
    type: object
  github_com_ai-comic-generator_server_internal_model.PublishResult:
    properties:
      articleUrl:
        description: 发布后的文章链接（若有）
        type: string
      mediaId:
        description: 微信素材库中的 media_id
        type: string
      platform:
        description: 发布平台，当前为微信公众号 WECHAT_MP
        example: WECHAT_MP
        type: string
      publishedAt:
        description: 发布时间（未发布时为 nil）
        type: string
      status:
        description: 发布状态：草稿 / 已发布 / 失败
        enum:
        - DRAFT
        - PUBLISHED
        - FAILED
        type: string
      title:
        description: 发布标题（来自故事构思）
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.QueryComicRequest:
    properties:
      pageNum:
        example: 1
        type: integer
      pageSize:
        example: 10
        type: integer
      phase:
        type: string
      status:
        enum:
        - PENDING
        - PROCESSING
        - AWAITING_CONFIRM
        - TITLE_CONFIRMED
        - AWAITING_STORYBOARD
        - COMPLETED
        - FAILED
        type: string
      userId:
        type: integer
    type: object
...
  github_com_ai-comic-generator_server_internal_model.RegeneratePanelRequest:
    properties:
      dialogue:
        items:
          type: string
        type: array
      narration:
        type: string
      panelNo:
        type: integer
      scene:
        type: string
      taskId:
        type: string
    required:
    - panelNo
    - taskId
    type: object
...
  github_com_ai-comic-generator_server_internal_model.RetryComicRequest:
    properties:
      taskId:
        type: string
    required:
    - taskId
    type: object
  github_com_ai-comic-generator_server_internal_model.StartComicRequest:
    properties:
      taskId:
        type: string
    required:
    - taskId
    type: object
...
  github_com_ai-comic-generator_server_internal_model.StoryIdeationResult:
    properties:
      highlights:
        description: 亮点情节列表
        items:
          type: string
        type: array
      keyConflict:
        description: 核心冲突（推动剧情的主矛盾）
        type: string
      synopsis:
        description: 故事梗概（一段话摘要）
        type: string
      theme:
        description: 主题（如成长、友情）
        type: string
      title:
        description: AI 推荐的漫画标题
        type: string
      tone:
        description: 基调（搞笑 / 热血 / 治愈等）
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.StoryboardPanel:
    properties:
      camera:
        description: 镜头类型（特写 / 中景 / 全景等）
        type: string
      dialogue:
        description: 角色台词列表（中文）
        items:
          type: string
        type: array
      dialogueEn:
        description: 台词英文版（供 bubble 模式 / 英文气泡使用）
        items:
          type: string
        type: array
      imagePrompt:
        description: 该格 AI 生图的英文 Prompt
        type: string
      narration:
        description: 旁白文字
        type: string
      panelNo:
        description: 分镜格序号，从 1 开始
        type: integer
      scene:
        description: 场景与画面描述
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.StoryboardResult:
    properties:
      pageCount:
        description: 预计页数
        type: integer
      panels:
        description: 全部分镜格列表
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.StoryboardPanel'
        type: array
    type: object
  github_com_ai-comic-generator_server_internal_model.TitleOption:
    properties:
      subtitle:
        description: 副标题/卖点，可选
        type: string
      title:
        description: 漫画标题，不超过 20 字
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.TitleOptionsResult:
    properties:
      options:
        description: 3-5 个标题方案
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.TitleOption'
        type: array
    type: object
...
info:
  contact: {}
  description: REST API。HTTP 多为 200，业务结果看 JSON 的 code（0 成功）；鉴权接口需 Cookie session。
  title: AI Comic Generator API
  version: "1.0"
paths:
  /comic/confirm-storyboard:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分镜列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ConfirmStoryboardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 确认分镜
      tags:
      - 漫画
  /comic/confirm-title:
    post:
      consumes:
      - application/json
      parameters:
      - description: taskId 与 title
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ConfirmTitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 确认标题
      tags:
      - 漫画
  /comic/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 主题与风格等
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.CreateComicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 创建漫画任务
      tags:
      - 漫画
...
  /comic/get:
    get:
      parameters:
      - description: 任务 ID
        in: query
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 查询漫画任务
      tags:
      - 漫画
  /comic/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页与筛选
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.QueryComicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ComicPageResult'
              type: object
      security:
      - SessionCookie: []
      summary: 漫画任务分页
      tags:
      - 漫画
  /comic/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: taskId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PublishComicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 发布到公众号
      tags:
      - 漫画
  /comic/regenerate-panel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 格号与场景
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.RegeneratePanelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 单格重绘
      tags:
      - 漫画
  /comic/retry:
    post:
      consumes:
      - application/json
      parameters:
      - description: taskId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.RetryComicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 失败重试
      tags:
      - 漫画
  /comic/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: taskId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.StartComicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 启动生成流水线
      tags:
      - 漫画
...
schemes:
- http
- https
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"