# server/internal/handler/comic_handler.go
package handler // HTTP 请求处理层：漫画任务相关接口

import (
	"net/http" // HTTP 状态码（统一返回 200，业务码在 JSON body 中）

	"github.com/ai-comic-generator/server/internal/common" // 统一响应封装、错误码、角色常量
	"github.com/ai-comic-generator/server/internal/middleware" // 从 Context 读取 AuthCheck 注入的登录用户
	"github.com/ai-comic-generator/server/internal/model"  // 请求/响应数据结构
	"github.com/ai-comic-generator/server/internal/service" // 漫画业务逻辑层
	"github.com/gin-gonic/gin" // Web 框架，提供 Context、路由、JSON 绑定
)

// ComicHandler 漫画任务 HTTP 入口
type ComicHandler struct {
	svc *service.ComicService // 漫画任务业务服务（创建、查询、分页）
}

// NewComicHandler 创建漫画 Handler，注入业务服务依赖
func NewComicHandler(svc *service.ComicService) *ComicHandler {
	return &ComicHandler{svc: svc} // 保存服务引用供各接口方法调用
}

// Create 创建漫画生成任务（异步标题推荐 + 五步创作流水线）
// @Summary      创建漫画任务
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.CreateComicRequest  true  "主题与风格等"
// @Success      200   {object}  common.BaseResponse
// @Security     SessionCookie
// @Router       /comic/create [post]
func (h *ComicHandler) Create(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c) // 从 AuthCheck 中间件注入的 Context 读取登录用户
	if !ok { // 未挂中间件或登录态异常（正常流程下 AuthCheck 已拦截）
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin)) // 防御性返回未登录错误
		return // 终止处理
	}

	var req model.CreateComicRequest                               // 声明创建请求体变量
	if err := c.ShouldBindJSON(&req); err != nil {                 // 将 JSON 请求体绑定到结构体并校验 required 字段
		c.JSON(http.StatusOK, common.Error(common.ErrParams)) // 参数缺失或格式错误
		return // 终止处理
	}

	taskID, err := h.svc.Create(loginUser.ID, &req) // 调用业务层创建任务，后台异步跑标题推荐
	if err != nil { // 创建失败（如数据库写入错误）
		handleError(c, err) // 将业务错误映射为统一 JSON 响应
		return // 终止处理
	}
	c.JSON(http.StatusOK, common.Success(gin.H{"taskId": taskID}))
}

// ConfirmTitle 用户确认标题（不启动流水线）
// @Summary      确认标题
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.ConfirmTitleRequest  true  "taskId 与 title"
// @Success      200   {object}  common.BaseResponse{data=bool}
// @Security     SessionCookie
// @Router       /comic/confirm-title [post]
func (h *ComicHandler) ConfirmTitle(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c)
	if !ok {
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin))
		return
	}

	var req model.ConfirmTitleRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusOK, common.Error(common.ErrParams))
		return
	}

	if err := h.svc.ConfirmTitle(loginUser.ID, &req, isAdminUser(loginUser)); err != nil {
		handleError(c, err)
		return
	}
	c.JSON(http.StatusOK, common.Success(true))
}

// Start 正式启动后续五步流水线
// @Summary      启动生成流水线
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.StartComicRequest  true  "taskId"
// @Success      200   {object}  common.BaseResponse{data=bool}
// @Security     SessionCookie
// @Router       /comic/start [post]
func (h *ComicHandler) Start(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c)
	if !ok {
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin))
		return
	}

	var req model.StartComicRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusOK, common.Error(common.ErrParams))
		return
	}

	if err := h.svc.StartPipeline(loginUser.ID, &req, isAdminUser(loginUser)); err != nil {
		handleError(c, err)
		return
	}
	c.JSON(http.StatusOK, common.Success(true))
}

// ConfirmStoryboard 确认/编辑分镜后启动生图
// @Summary      确认分镜
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.ConfirmStoryboardRequest  true  "分镜列表"
// @Success      200   {object}  common.BaseResponse{data=bool}
// @Security     SessionCookie
// @Router       /comic/confirm-storyboard [post]
func (h *ComicHandler) ConfirmStoryboard(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c)
	if !ok {
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin))
		return
	}

	var req model.ConfirmStoryboardRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusOK, common.Error(common.ErrParams))
		return
	}

	if err := h.svc.ConfirmStoryboard(loginUser.ID, &req, isAdminUser(loginUser)); err != nil {
		handleError(c, err)
		return
	}
	c.JSON(http.StatusOK, common.Success(true))
}

// Retry 失败任务从当前步骤重试
// @Summary      失败重试
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.RetryComicRequest  true  "taskId"
// @Success      200   {object}  common.BaseResponse{data=bool}
// @Security     SessionCookie
// @Router       /comic/retry [post]
func (h *ComicHandler) Retry(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c)
	if !ok {
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin))
		return
	}

	var req model.RetryComicRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusOK, common.Error(common.ErrParams))
		return
	}

	if err := h.svc.RetryFailed(loginUser.ID, &req, isAdminUser(loginUser)); err != nil {
		handleError(c, err)
		return
	}
	c.JSON(http.StatusOK, common.Success(true))
}

// RegeneratePanel 修改单格后重绘并重排版
// @Summary      单格重绘
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.RegeneratePanelRequest  true  "格号与场景"
// @Success      200   {object}  common.BaseResponse
// @Security     SessionCookie
// @Router       /comic/regenerate-panel [post]
func (h *ComicHandler) RegeneratePanel(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c)
	if !ok {
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin))
		return
	}

	var req model.RegeneratePanelRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusOK, common.Error(common.ErrParams))
		return
	}

	info, err := h.svc.RegeneratePanel(loginUser.ID, &req, isAdminUser(loginUser))
	if err != nil {
		handleError(c, err)
		return
	}
	c.JSON(http.StatusOK, common.Success(info))
}

// Publish 将已完成作品发布至微信公众号
// @Summary      发布到公众号
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.PublishComicRequest  true  "taskId"
// @Success      200   {object}  common.BaseResponse
// @Security     SessionCookie
// @Router       /comic/publish [post]
func (h *ComicHandler) Publish(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c)
	if !ok {
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin))
		return
	}

	var req model.PublishComicRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusOK, common.Error(common.ErrParams))
		return
	}

	result, err := h.svc.Publish(loginUser.ID, &req, isAdminUser(loginUser))
	if err != nil {
		handleError(c, err)
		return
	}
	c.JSON(http.StatusOK, common.Success(result))
}

// Get 查询任务详情
// @Summary      查询漫画任务
// @Tags         漫画
// @Produce      json
// @Param        taskId  query  string  true  "任务 ID"
// @Success      200     {object}  common.BaseResponse
// @Security     SessionCookie
// @Router       /comic/get [get]
func (h *ComicHandler) Get(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c) // 从 AuthCheck 中间件注入的 Context 读取登录用户
	if !ok { // 未挂中间件或登录态异常
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin)) // 防御性返回未登录错误
		return // 终止处理
	}
	taskID := c.Query("taskId") // 从 URL 查询参数获取任务 ID
	if taskID == "" { // 未传 taskId
		c.JSON(http.StatusOK, common.Error(common.ErrParams)) // 返回参数错误
		return // 终止处理
	}

	isAdmin := isAdminUser(loginUser)                              // 根据用户角色判断是否管理员
	info, err := h.svc.GetForUser(taskID, loginUser.ID, isAdmin) // 查询详情；非管理员只能看自己的任务
	if err != nil { // 任务不存在或无权限
		handleError(c, err) // 映射为 NOT_FOUND / NO_AUTH 等错误响应
		return // 终止处理
	}
	c.JSON(http.StatusOK, common.Success(info)) // 返回漫画任务详情（含各步产物 JSON）
}

// ListPage 分页查询漫画任务
// @Summary      漫画任务分页
// @Tags         漫画
// @Accept       json
// @Produce      json
// @Param        body  body  model.QueryComicRequest  true  "分页与筛选"
// @Success      200   {object}  common.BaseResponse{data=model.ComicPageResult}
// @Security     SessionCookie
// @Router       /comic/page [post]
func (h *ComicHandler) ListPage(c *gin.Context) {
	loginUser, ok := middleware.GetLoginUserFromContext(c) // 从 AuthCheck 中间件注入的 Context 读取登录用户
	if !ok { // 未挂中间件或登录态异常
		c.JSON(http.StatusOK, common.Error(common.ErrNotLogin)) // 防御性返回未登录错误
		return // 终止处理
	}

	var req model.QueryComicRequest                        // 声明分页查询请求体
	if err := c.ShouldBindJSON(&req); err != nil {         // 绑定 JSON（pageNum、pageSize、筛选条件等）
		c.JSON(http.StatusOK, common.Error(common.ErrParams)) // 参数格式错误
		return // 终止处理
	}

	if !isAdminUser(loginUser) { // 普通用户不能查他人任务
		req.UserID = &loginUser.ID // 强制限定为当前用户 ID
	}

	page, err := h.svc.ListByPage(&req) // 调用业务层分页查询
	if err != nil { // 数据库查询失败
		handleError(c, err) // 映射为系统错误响应
		return // 终止处理
	}
	c.JSON(http.StatusOK, common.Success(page)) // 返回分页结果（total、records、pageNum、pageSize）
}

// isAdminUser 根据中间件已加载的用户对象判断是否为管理员（无需再次查库）
func isAdminUser(user *model.User) bool {
	return user.UserRole == common.AdminRole // 角色等于 admin 则为管理员
}